TIMELINE PAGE AND AUTH FUNCTIONALITY UPDATED (#153)
* Auth Changed 

Auth is no longer required to be a page hence almost everything is changing here in terms of code structure..
The modal code is added here and then this page is included at the bottom of the home page so when the 
login button is clicked at the home page, the content of this page loads..

Therefore, there is no need for it's own head links and the likes.. all connection it depends on would be placed in the home page.

* Internal CSS to header  …

Header updated with internal css for timeline.php

* Timeline Page Optimized  …

This page was assigned to me for optimization for both responsiveness and better presentation still according to approved design. Do not be afraid to merge.

* Updated to include modal functionality

Now the button no longer functions as a link to another page but as a trigger for the login modal.

The button code is changed and the auth.php file is included at the bottom of this page..

Thanks.